chore: fix file name
Missed pushing this fix in the latest PR.
